+++ ecc/scripts/constants.js
